git commit total 출력

+++ Backend/src/main/java/com/pg/programmerground/dto/GithubTotalDto.java
@@ -17,7 +17,7 @@ public class GithubTotalDto {
     private Long totalRepo;
 
     @JsonProperty("total_commit")
-    private Long totalCommit;
+    private Integer totalCommit;
 
     @JsonProperty("total_star")
     private Long totalStar;
